pyproject.toml
[tool.poetry]
name = "to-do-list"
version = "0.1.0"
description = ""
authors = ["Bruno Carvalho da Silva <bruno.silva@lsdi.ufma.br>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = {extras = ["standard"], version = "^0.115.6"}
sqlalchemy = "^2.0.36"
pydantic-settings = "^2.7.1"
alembic = "^1.14.0"
pytz = "^2024.2"


[tool.poetry.group.dev.dependencies]
pytest-cov = "^6.0.0"
taskipy = "^1.14.1"
ruff = "^0.8.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
pythonpath = "."
addopts = '-p no:warnings'

[tool.taskipy.tasks]
run = 'fastapi dev to_do_list/app.py'
test = 'pytest -s -x --cov=to_do_list -vv'